Crate zkp_server

zkp_server includes modules and functions that zkp_client communicates with.

Functionalities such as:

Serving requests on top of TCP protocol using JSON for data exchange and simplicity.

Registering the user’s commits using an in-memory session store

Presenting Authentication challenge as part of the login flow

Validating the answer from client before providing access

Constraints

Following Constraints are set for convenience

// set so that `s = k - c.x` calculation doesn't result in a negative value.
// see constraints for zkp_client and zkp_common for more details
pub const MAX_CHALLENGE_VAL: u32 = 4;

Session Store

Below structure represents the in-memory session store used to store user sessions

#[derive(Debug)]
pub struct User {
    pub username: String,
    pub commits: Commits,
    pub is_verified: bool,
    pub challenge: Option<Challenge>,
    pub last_verified: Option<DateTime<Utc>>,
    pub last_login: Option<DateTime<Utc>>,
}
 
pub struct SessionStore {
    pub users: HashMap<Username, User>,
}

Performance Improvements

Server performance can be further improved by adopting a multi-threaded model or by implementing async/await pattern using runtimes such as Tokio

The Session storage would also need to be enhanced when using a multi-threaded model.

Either using Arc Mutex for shared-state concurrency or using a dedicated, external key/value store for session management.

JSON is used for simplicity, but using a binary based representation such as protocol buffers for data exchange can further increase serialization and deserialization performance.

Modules

challenge

handlers

session_store

Constants

MAX_CHALLENGE_VAL

Functions

init_logger

init_session_store